refactor(coverage): 共享实时完整性判定
